pyproject.toml
[tool.poetry]
name = "farmrpg-etl"
version = "0.1.0"
description = ""
authors = ["Noah Kantrowitz <noah@coderanger.net>"]
license = "Apache-2.0"
readme = "README.md"
packages = [{include = "farmrpg_etl", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
httpx = "^0.22.0"
attrs = "^21.4.0"
beautifulsoup4 = "^4.11.1"
lxml = "^4.8.0"
structlog = "^21.5.0"
google-cloud-firestore = "^2.4.0"
cattrs = "^22.1.0"
exceptiongroup = "^1.0.0-rc.3"
orm = {git = "https://github.com/encode/orm.git", extras=["postgresql"]}
starlette = "^0.20.0"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
firebase-admin = "^5.2.0"
google-cloud-iam = "^2.6.1"
typer = "^0.4.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.1.1"
freezegun = "^1.2.1"
black = "^22.3.0"
pytest-asyncio = "^0.18.3"


[tool.poetry.group.migrations.dependencies]
alembic = "^1.7.7"
psycopg2 = "^2.9.3"

[tool.isort]
known_first_party= ["farmrpg_etl"]

[tool.pytest.ini_options]
asyncio_mode = "strict"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
